schema,boston_data,evidentli_data_drift_detection,data_validation,model_trainer,artfact_entity,training_pipeline,main_utils

+++ YOUTH_MIGRATION/utils/main_utils.py
@@ -6,102 +6,111 @@ import dill
 import yaml
 from pandas import DataFrame
 
-from us_visa.exception import USvisaException
-from us_visa.logger import logging
+from youth_migration.exception import CustomException
+from youth_migration.logger import logging
 
 
+# -------------------------------
+# YAML READ
+# -------------------------------
 def read_yaml_file(file_path: str) -> dict:
     try:
         with open(file_path, "rb") as yaml_file:
             return yaml.safe_load(yaml_file)
 
     except Exception as e:
-        raise USvisaException(e, sys) from e
+        raise CustomException(e, sys) from e
 
 
+# -------------------------------
+# YAML WRITE
+# -------------------------------
 def write_yaml_file(file_path: str, content: object, replace: bool = False) -> None:
     try:
-        if replace:
-            if os.path.exists(file_path):
-                os.remove(file_path)
+        if replace and os.path.exists(file_path):
+            os.remove(file_path)
+
         os.makedirs(os.path.dirname(file_path), exist_ok=True)
+
         with open(file_path, "w") as file:
             yaml.dump(content, file)
+
     except Exception as e:
-        raise USvisaException(e, sys) from e
+        raise CustomException(e, sys) from e
 
 
+# -------------------------------
+# LOAD OBJECT
+# -------------------------------
 def load_object(file_path: str) -> object:
-    logging.info("Entered the load_object method of utils")
+    logging.info("Entered load_object method")
 
     try:
-
         with open(file_path, "rb") as file_obj:
             obj = dill.load(file_obj)
 
-        logging.info("Exited the load_object method of utils")
-
+        logging.info("Exited load_object method")
         return obj
 
     except Exception as e:
-        raise USvisaException(e, sys) from e
+        raise CustomException(e, sys) from e
 
+
+# -------------------------------
+# SAVE NUMPY ARRAY
+# -------------------------------
 def save_numpy_array_data(file_path: str, array: np.array):
-    """
-    Save numpy array data to file
-    file_path: str location of file to save
-    array: np.array data to save
-    """
     try:
-        dir_path = os.path.dirname(file_path)
-        os.makedirs(dir_path, exist_ok=True)
+        os.makedirs(os.path.dirname(file_path), exist_ok=True)
+
         with open(file_path, 'wb') as file_obj:
             np.save(file_obj, array)
+
     except Exception as e:
-        raise USvisaException(e, sys) from e
+        raise CustomException(e, sys) from e
 
 
+# -------------------------------
+# LOAD NUMPY ARRAY
+# -------------------------------
 def load_numpy_array_data(file_path: str) -> np.array:
-    """
-    load numpy array data from file
-    file_path: str location of file to load
-    return: np.array data loaded
-    """
     try:
         with open(file_path, 'rb') as file_obj:
             return np.load(file_obj)
+
     except Exception as e:
-        raise USvisaException(e, sys) from e
+        raise CustomException(e, sys) from e
 
 
+# -------------------------------
+# SAVE OBJECT
+# -------------------------------
 def save_object(file_path: str, obj: object) -> None:
-    logging.info("Entered the save_object method of utils")
+    logging.info("Entered save_object method")
 
     try:
         os.makedirs(os.path.dirname(file_path), exist_ok=True)
+
         with open(file_path, "wb") as file_obj:
             dill.dump(obj, file_obj)
 
-        logging.info("Exited the save_object method of utils")
+        logging.info("Exited save_object method")
 
     except Exception as e:
-        raise USvisaException(e, sys) from e
+        raise CustomException(e, sys) from e
 
 
-def drop_columns(df: DataFrame, cols: list)-> DataFrame:
-
-    """
-    drop the columns form a pandas DataFrame
-    df: pandas DataFrame
-    cols: list of columns to be dropped
-    """
-    logging.info("Entered drop_columns methon of utils")
+# -------------------------------
+# DROP COLUMNS
+# -------------------------------
+def drop_columns(df: DataFrame, cols: list) -> DataFrame:
+    logging.info("Entered drop_columns method")
 
     try:
         df = df.drop(columns=cols, axis=1)
 
-        logging.info("Exited the drop_columns method of utils")
-        
+        logging.info("Exited drop_columns method")
         return df
+
     except Exception as e:
-        raise USvisaException(e, sys) from e
+        raise CustomException(e, sys) from e
